requirements.txt
sanic~=21.9.1
PyJWT~=2.2.0